Bugs fixes in loss function W_CC_Loss

+++ segmentation/losses.py
@@ -374,7 +374,7 @@ def Weighted_CatCross_Loss(numClasses):
 
         SDM = tf.py_function(func=calc_dist_map_batch,
                              inp=[y_true],
-                             Tout=tf.float32)
+                             Tout=tf.float64)
 
         epsilon = backend_config.epsilon
         gamma = 8
